requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
lxml==4.9.2
moviepy==1.0.3
numpy==1.25.0
Pillow==9.5.0
proglog==0.1.10
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
